Added additional camera functionality
If a player move too far away from the group and cause the camera to move too far out, that player will now break away from the camera's group tracking, and the camera will refocus on the main group. 

This is a functionality that will support future breakaway cameras for player that move too far away from the main group.